Refactor: pages 디렉토리 잘못된 폴더구조 교체
features 내부 디렉토리가 pages 내부로 들어가 있는 문제 해결
