+++ src/app/router.tsx
@@ -1,8 +1,8 @@
-import DashboardPage from "@/pages/dashboard/DashboardPage";
-import HomePage from "@/pages/home/HomePage";
-import OverviewPage from "@/pages/overview/OverviewPage";
-import StockDetailPage from "@/pages/stockDetail/StockDetailPage";
-import Test from "@/pages/test/Test";
+import DashboardPage from "@/pages/DashboardPage";
+import HomePage from "@/pages/HomePage";
+import OverviewPage from "@/pages/OverviewPage";
+import StockDetailPage from "@/pages/StockDetailPage";
+import TestPage from "@/pages/TestPage";
 
 import { createBrowserRouter } from "react-router-dom";
 
@@ -25,7 +25,7 @@ const router = createBrowserRouter([
   },
   {
     path: "/test",
-    element: <Test />,
+    element: <TestPage />,
   },
 ]);
