+++ src/pages/StockDetailPage.tsx
